Hi all!
The project "squeak-app" is part of the continuous build infrastructure (short: CI) for Squeak's bundles, which you can download at https://files.squeak.org/, followed by the version tag you are looking for.

All bundles are updated on a regular basis, which is daily for Trunk (i.e. the current alpha version) and monthly for release versions.

Note that there won't be new bundles if Squeak's build number, which reflects in-image code updates, did not change between CI jobs.

-- The Squeak Oversight Board

[ci skip]