Fix squeak.sh command line for running script files
Since 0724852c3dc95eef554eb9bce9c8bc6da3a919c8, it was no longer possible to pass a single script file to the squeak.sh file in the AiO bundle like this:

	./squeak.sh /path/to/script.st

(Apparently, this was also never possible in the linux bundle.) With this patch, this is made possible for AiO and Linux bundles (again).

Note that the current argument parser in squeak.sh is limited with regard to the IMAGE and STARGS arguments: Image files are only accepted when the extension is .image and script files are only accepted when their extension is .st or .cs.